+++ Source/MusicServer/Genre.m
@@ -66,7 +66,7 @@ NSString *gUnknownGenreName = @"gUnknownGenreName";
 	self = [super init];
 	mName = [[decoder decodeObject] retain];
 	
-	unsigned length;
+	NSUInteger length;
 	memcpy(&_identifier, [decoder decodeBytesWithReturnedLength:&length], sizeof(_identifier));
 	
 	return self;
