+++ Source/MusicServer/Song.m
@@ -408,7 +408,7 @@ NSString *gSongPropertyIsCompilation = @"gSongPropertyIsCompilation";
 	self = [super init];
 	mValues = [[decoder decodeObject] retain];
 	
-	unsigned length;
+	NSUInteger length;
 	memcpy(&mValid, [decoder decodeBytesWithReturnedLength:&length], sizeof(mValid));
 	
 	return self;
